update: 9/1/2024

+++ app/Views/master/kegiatan_mitra/detail.php
@@ -47,67 +47,90 @@
                 <div class="card">
                     <div class="card-header p-2">
                         <ul class="nav nav-pills">
-                            <li class="nav-item"><a class="nav-link active" href="#settings" data-toggle="tab">Tambah Data</a></li>
-                            <li class="nav-item"><a class="nav-link " href="#activity" data-toggle="tab">Mitra Lapangan</a></li>
+                            <?php if (in_groups('admin')) { ?>
+                                <li class="nav-item"><a class="nav-link <?= in_groups('admin') ? 'active' : '' ?>" href="#settings" data-toggle="tab">Tambah Data</a></li>
+                            <?php } ?>
+                            <li class="nav-item"><a class="nav-link <?= in_groups('kepala') ? 'active' : '' ?> " href="#activity" data-toggle="tab">Mitra Lapangan</a></li>
                             <li class="nav-item"><a class="nav-link" href="#timeline" data-toggle="tab">Mitra Pengolahan</a></li>
+
                         </ul>
                     </div>
                     <div class="card-body">
+
                         <div class="tab-content">
-
-                            <div class="tab-pane " id="activity">
-
-                                <table id="example2" class="table table-bordered table-hover">
+                            <div class="tab-pane  <?= in_groups('kepala') ? 'active' : '' ?>" id="activity">
+                                <?php if (!empty($lapangan)) { ?>
+                                    <a href="<?= base_url() ?>kegiatan_mitra/cetak_mitra_lapangan/<?= $kegiatan['id_kegiatan'] ?>" disabled class="btn btn-success" target="_blank"><i class="fas fa-save"></i> Cetak</a>
+                                <?php  } else { ?>
+                                    <button class="btn btn-secondary" target="_blank"><i class="fas fa-save"></i> Cetak</button>
+                                <?php  } ?>
+                                <table class="table table-bordered table-hover">
                                     <thead>
                                         <tr>
                                             <th>No</th>
                                             <th>Nik</th>
                                             <th>Nama</th>
-
-                                            <th>Aksi</th>
+                                            <?php if (in_groups('admin')) { ?>
+                                                <th>Aksi</th>
+                                            <?php } ?>
                                         </tr>
                                     </thead>
                                     <tbody>
                                         <?php $i = 1 ?>
+                                        <?php if (empty($lapangan)) { ?>
+                                            <tr>
+                                                <td colspan="5" class="text-center">Data Masih Kosong</td>
+                                            </tr>
+                                        <?php  } ?>
                                         <?php foreach ($lapangan as $l) { ?>
                                             <tr>
                                                 <td><?= $i++ ?></td>
                                                 <td><?= $l['nik'] ?></td>
                                                 <td><?= $l['nama_mitra'] ?></td>
-
-
-                                                <td><a href="<?= base_url() ?>kegiatan_mitra/hapus/<?= $l['id_kegiatan_mitra'] ?>/ <?= $kegiatan['id_kegiatan'] ?>" class="btn btn-danger" onclick="return confirm('Apakah Anda yakin untuk menghapus mitra ini?')">Hapus</a>
-                                                </td>
+                                                <?php if (in_groups('admin')) { ?>
+                                                    <td><a href="<?= base_url() ?>kegiatan_mitra/hapus/<?= $l['id_kegiatan_mitra'] ?>/ <?= $kegiatan['id_kegiatan'] ?>" class="btn btn-danger" onclick="return confirm('Apakah Anda yakin untuk menghapus mitra ini?')">Hapus</a>
+                                                    </td>
+                                                <?php } ?>
                                             </tr>
                                         <?php  } ?>
-
                                     </tbody>
                                 </table>
-
                             </div>
 
                             <div class="tab-pane" id="timeline">
-                                <table id="example1" class="table table-bordered table-hover">
+                                <?php if (!empty($pengolahan)) { ?>
+                                    <a href="<?= base_url() ?>kegiatan_mitra/cetak_mitra_pengolahan/<?= $kegiatan['id_kegiatan'] ?>" disabled class="btn btn-success" target="_blank"><i class="fas fa-save"></i> Cetak</a>
+                                <?php  } else { ?>
+                                    <button class="btn btn-secondary" target="_blank"><i class="fas fa-save"></i> Cetak</button>
+                                <?php  } ?>
+                                <table class="table table-bordered table-hover">
                                     <thead>
                                         <tr>
                                             <th>No</th>
                                             <th>Nik</th>
                                             <th>Nama</th>
-
-                                            <th>Aksi</th>
+                                            <?php if (in_groups('admin')) { ?>
+                                                <th>Aksi</th>
+                                            <?php } ?>
                                         </tr>
                                     </thead>
                                     <tbody>
                                         <?php $i = 1 ?>
+                                        <?php if (empty($pengolahan)) { ?>
+                                            <tr>
+                                                <td colspan="5" class="text-center">Data Masih Kosong</td>
+                                            </tr>
+                                        <?php  } ?>
                                         <?php foreach ($pengolahan as $p) { ?>
                                             <tr>
                                                 <td><?= $i++ ?></td>
                                                 <td><?= $p['nik'] ?></td>
                                                 <td><?= $p['nama_mitra'] ?></td>
 
-
-                                                <td><a href="<?= base_url() ?>kegiatan_mitra/hapus/<?= $p['id_kegiatan_mitra'] ?>" class="btn btn-danger" onclick="return confirm('Apakah Anda yakin untuk menghapus mitra ini?')">Hapus</a>
-                                                </td>
+                                                <?php if (in_groups('admin')) { ?>
+                                                    <td><a href="<?= base_url() ?>kegiatan_mitra/hapus/<?= $p['id_kegiatan_mitra'] ?>" class="btn btn-danger" onclick="return confirm('Apakah Anda yakin untuk menghapus mitra ini?')">Hapus</a>
+                                                    </td>
+                                                <?php } ?>
                                             </tr>
                                         <?php  } ?>
 
@@ -116,7 +139,7 @@
 
                             </div>
 
-                            <div class="tab-pane active" id="settings">
+                            <div class="tab-pane <?= in_groups('admin') ? 'active' : '' ?>" id="settings">
                                 <form class="form-horizontal" action="<?= base_url() ?>kegiatan_mitra/simpan/<?= $kegiatan['id_kegiatan'] ?>" method="post">
                                     <div class="form-group row">
                                         <label for="inputName" class="col-sm-2 col-form-label">NIK</label>
